using doubles again for use with mi library

+++ read_binary.m
@@ -3,7 +3,9 @@ function [bin_import, spp] = read_binary()
 fid = fopen('samplebuffer_sanmiguel_diffuse_8spp.bin');
 spp = 8;
 
-bin_import = single(fread (fid, [9+3*10, 362*620*8],'float'));
+% bin_import = single(fread (fid, [9+3*10, 362*620*8],'float'));
+% mutual information library only takes double
+bin_import = fread (fid, [9+3*10, 362*620*8],'float');
 'finished reading'
 
 fclose(fid);
